Added "version" subcommand
