+++ wordfence/cli/version/version.py
@@ -1,0 +1,11 @@
+from ..subcommands import Subcommand
+
+
+class VersionSubcommand(Subcommand):
+
+    def invoke(self) -> int:
+        self.context.display_version()
+        return 0
+
+
+factory = VersionSubcommand
